slim down template - move import + export logic to settings.mjs
